+++ src/Math/float3x4.cpp
@@ -1102,7 +1102,7 @@ void float3x4::BatchTransformPos(float3 *pointArray, int numPoints, int stride) 
 	if (!pointArray)
 		return;
 #endif
-	assume(stride >= sizeof(float3));
+	assume(stride >= (int)sizeof(float3));
 	u8 *data = reinterpret_cast<u8*>(pointArray);
 	for(int i = 0; i < numPoints; ++i)
 	{
@@ -1129,7 +1129,7 @@ void float3x4::BatchTransformDir(float3 *dirArray, int numVectors, int stride) c
 	if (!dirArray)
 		return;
 #endif
-	assume(stride >= sizeof(float3));
+	assume(stride >= (int)sizeof(float3));
 	u8 *data = reinterpret_cast<u8*>(dirArray);
 	for(int i = 0; i < numVectors; ++i)
 	{
@@ -1156,7 +1156,7 @@ void float3x4::BatchTransform(float4 *vectorArray, int numVectors, int stride) c
 	if (!vectorArray)
 		return;
 #endif
-	assume(stride >= sizeof(float4));
+	assume(stride >= (int)sizeof(float4));
 	u8 *data = reinterpret_cast<u8*>(vectorArray);
 	for(int i = 0; i < numVectors; ++i)
 	{
